/product/getCateProperty

+++ product/addV2.go
@@ -54,19 +54,19 @@ type QualityAttachments struct {
 
 //返回参数
 type AddV2 struct {
-	ProductId      int    `json:"product_id"`
-	OutProductId   int    `json:"out_product_id"`
+	ProductId      int64  `json:"product_id"`
+	OutProductId   int64  `json:"out_product_id"`
 	OuterProductId string `json:"outer_product_id"`
 	CreateTime     string `json:"create_time"`
 	Sku            []*Sku `json:"sku"`
 }
 
 type Sku struct {
-	SkuId         int    `json:"sku_id"`
-	OutSkuId      int    `json:"out_sku_id"`
+	SkuId         int64  `json:"sku_id"`
+	OutSkuId      int64  `json:"out_sku_id"`
 	OuterSkuId    string `json:"outer_sku_id"`
 	Code          string `json:"create_time"`
-	SpecDetailId1 int    `json:"spec_detail_id1"`
-	SpecDetailId2 int    `json:"spec_detail_id2"`
-	SpecDetailId3 int    `json:"spec_detail_id3"`
+	SpecDetailId1 int64  `json:"spec_detail_id1"`
+	SpecDetailId2 int64  `json:"spec_detail_id2"`
+	SpecDetailId3 int64  `json:"spec_detail_id3"`
 }
